fixed call to tp_2symbol to satisfy CRAY.
svn:revision:10223
cvs:account:jra
cvs:branch:trunk

+++ libbn/number.c
@@ -44,7 +44,7 @@ int	digits;		/* # digits wide */
 		digits = 1;
 	sprintf( fmt, "%%%dg", digits, digits );
 	sprintf( buf, fmt, input );
-	tp_2symbol( fp, buf, x, y, cscale, theta );
+	tp_2symbol( fp, buf, (double)x, (double)y, (double)cscale, theta );
 }
 
 void
